feat(capitelize function and test)
The capitalize function takes a string as parameter to then capitalize the first character of the string and removes the origin first character to replace it with the new capitalize one. The capitalize test expects that a given array with no capital letters will return a new one with the first letter being capitalized.