feat(caesarCipher function)
The function has two parameters, a string and a key to shift that string.

The first conditional statement is a base case which checks if wether the key is less than 0 or greater than 25, stopping the function. Then two variables are created, result wich has an empty value. The second one is alphabeth wich contains an empty array. Right afert that, a for loop is created that will populate the alphabet array with strings using the charCode method, iterating from the 97 to the 122 which are the codes for a to z in lowercase. Another loop is created after that which iterates in the given string, adding the result to the char variable after each iteration. While the char value is greater than a and lesser than z two other variables will be created, currentIndex which gets the index number within the alphabet array of the char variable. and a newIndex which uses that previous index and adds the key number or shift number, diving by 26 which are the available characters. The result will be add to the result variable, getting the string from alphabet using the newIndex value. A final else statment will handle the case when the char is not in lowercase which will return it without encrypt it.